requirements.txt
python-barcode==0.14.0
pyqrcode==1.2.1
flask==2.2.2
environs==9.5.0
pypng==0.20220715.0
Pillow==9.4.0
gunicorn==20.1.0